+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Change prescribed drugs toolbar title to `Medicines`
 - Fix downloaded prescription background color
 - Change UI for adding protocol/custom drugs
+- Show toolbar progress indicator in facility picker screen
 
 ### Fixes
 - Fix App crash when returning to the register patient screen before completing registration when patient date of birth is entered